linux-gen: init: split odp_internal.h
odp_internal.h contained code for three different purposes. Split
file into three different headers, so that it's no more a place
for random type/function prototype additions.

New files are:
* odp_init_internal.h which holds global/local init function
  prototypes
* odp_global_data.h which defines global data structure
* odp_sysinfo_internal.h which contains system info internal
  function prototypes
