+++ platform/linux-generic/include/odp_packet_io_ipc_internal.h
@@ -8,7 +8,6 @@
 #include <odp_packet_io_internal.h>
 #include <odp/api/packet.h>
 #include <odp_packet_internal.h>
-#include <odp_internal.h>
 #include <odp/api/shared_memory.h>
 
 #include <string.h>
