+++ platform/linux-generic/odp_crypto.c
@@ -8,7 +8,7 @@
 
 #include <odp_posix_extensions.h>
 #include <odp/api/crypto.h>
-#include <odp_internal.h>
+#include <odp_init_internal.h>
 #include <odp/api/spinlock.h>
 #include <odp/api/sync.h>
 #include <odp/api/debug.h>
