+++ platform/linux-generic/odp_ipsec_sad.c
@@ -12,6 +12,7 @@
 #include <odp/api/random.h>
 #include <odp/api/shared_memory.h>
 
+#include <odp_init_internal.h>
 #include <odp_debug_internal.h>
 #include <odp_ipsec_internal.h>
