+++ platform/linux-generic/odp_pool.c
@@ -14,7 +14,7 @@
 #include <odp/api/plat/thread_inlines.h>
 
 #include <odp_pool_internal.h>
-#include <odp_internal.h>
+#include <odp_init_internal.h>
 #include <odp_packet_internal.h>
 #include <odp_config_internal.h>
 #include <odp_debug_internal.h>
