+++ platform/linux-generic/odp_schedule_if.c
@@ -7,7 +7,7 @@
 #include "config.h"
 
 #include <odp_schedule_if.h>
-#include <odp_internal.h>
+#include <odp_init_internal.h>
 #include <odp_debug_internal.h>
 
 #include <stdlib.h>
