+++ platform/linux-generic/odp_thread.c
@@ -11,8 +11,8 @@
 #include <sched.h>
 #include <odp/api/thread.h>
 #include <odp/api/thrmask.h>
-#include <odp_internal.h>
 #include <odp/api/spinlock.h>
+#include <odp_init_internal.h>
 #include <odp_config_internal.h>
 #include <odp_debug_internal.h>
 #include <odp/api/shared_memory.h>
