+++ platform/linux-generic/odp_traffic_mngr.c
@@ -29,6 +29,8 @@
 #include <odp/api/plat/packet_inlines.h>
 #include <odp/api/plat/byteorder_inlines.h>
 #include <odp_macros_internal.h>
+#include <odp_init_internal.h>
+#include <odp_errno_define.h>
 
 /* Local vars */
 static const
