+++ platform/linux-generic/pktio/loop.c
@@ -13,6 +13,7 @@
 #include <odp_classification_internal.h>
 #include <odp_ipsec_internal.h>
 #include <odp_debug_internal.h>
+#include <odp_errno_define.h>
 #include <odp/api/plat/packet_flag_inlines.h>
 #include <odp/api/hints.h>
 #include <odp/api/plat/byteorder_inlines.h>
