+++ platform/linux-generic/pktio/sysfs.c
@@ -8,6 +8,7 @@
 
 #include <odp_api.h>
 #include <odp_packet_io_internal.h>
+#include <odp_errno_define.h>
 #include <errno.h>
 #include <string.h>
 #include <inttypes.h>
