+++ platform/linux-generic/pktio/tap.c
@@ -47,6 +47,7 @@
 #include <odp_packet_internal.h>
 #include <odp_packet_io_internal.h>
 #include <odp_classification_internal.h>
+#include <odp_errno_define.h>
 
 #define BUF_SIZE 65536
